+++ src/main/java/de/maxhenkel/car/config/ServerConfig.java
@@ -7,6 +7,7 @@ import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.material.Fluid;
 import net.neoforged.fml.event.config.ModConfigEvent;
 import net.neoforged.neoforge.common.ModConfigSpec;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -192,8 +193,18 @@ public class ServerConfig extends ConfigBase {
     }
 
     @Override
-    public void onReload(ModConfigEvent event) {
+    public void onReload(ModConfigEvent.Reloading event) {
         super.onReload(event);
+        onConfigChanged();
+    }
+
+    @Override
+    public void onLoad(ModConfigEvent.Loading evt) {
+        super.onLoad(evt);
+        onConfigChanged();
+    }
+
+    private void onConfigChanged() {
         gasStationValidFuelList = gasStationValidFuels.get().stream().map(TagUtils::getFluid).filter(Objects::nonNull).collect(Collectors.toList());
         generatorValidFuelList = generatorValidFuels.get().stream().map(TagUtils::getFluid).filter(Objects::nonNull).collect(Collectors.toList());
         carDriveBlockList = carDriveBlocks.get().stream().map(TagUtils::getBlock).filter(Objects::nonNull).collect(Collectors.toList());